design: 아침 대화 페이지 구현
